Refactor Maintenance and Tenant Services for Enhanced Data Management
- Introduced a new ReviewType enum to standardize review categorization across the application, improving data consistency.
- Refactored PropertyService to ensure accurate radius calculations for property queries, enhancing search functionality.
- Updated TenantService to utilize the new ReviewType enum for tenant reviews, streamlining review handling.
- Implemented a MaintenanceUniversalTableProvider for efficient management of maintenance issues, including pagination and filtering capabilities.
- Removed obsolete financial report components and integrated universal table structures for improved reporting consistency.
- Enhanced tenant management screens to utilize universal tables, simplifying data presentation and interaction.

+++ eRents.Application/Service/PropertyService/PropertyService.cs
@@ -593,7 +593,7 @@ namespace eRents.Application.Service.PropertyService
 
 				query = query.Where(p => p.Address != null && 
 					p.Address.Latitude.HasValue && p.Address.Longitude.HasValue &&
-					Math.Sqrt(Math.Pow((double)(p.Address.Latitude - lat), 2) + Math.Pow((double)(p.Address.Longitude - lng), 2)) <= radius);
+					Math.Sqrt(Math.Pow((double)(p.Address.Latitude - lat), 2) + Math.Pow((double)(p.Address.Longitude - lng), 2)) <= (double)radius);
 			}
 
 			return query;
